requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
redis
structlog
apscheduler
httpx
pandas
numpy
scikit-learn
xgboost
joblib
shap
sse-starlette
